+++ core/src/io.rs
@@ -1,3 +1,4 @@
 pub mod archived_queens;
+pub mod image;
 pub mod json;
 pub mod text;